(#60) - test in saucelabs and node
Swaps out Zuul for Beefy, so that we can test
in Saucelabs. Also adds Node tests, which actually
work because we are now using humble-localstorage
as a wrapper (1KB min+gz) for window.localStorage,
with the added benefit this this module will now
work in Safari private browsing (using an in-memory
fallback).