(#76) - fix localstorage destroy()
This fixes the PouchDB test suite. I did not add a new test
to this repo, because I didn't have time.

These tests were broken in Firefox and Safari but not Chrome.
Apparently the issue is that LocalStorage implementations are
allowed to reorder the keys before and after removing items,
meaning we can't count on the order to stay the same while
we're deleting stuff. The W3C spec has more info:
https://www.w3.org/TR/webstorage/